# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TribeHub API
  version: 0.1.0
  description: >
    TribeHub social network REST API.
    This is a placeholder spec — endpoints will be added as modules are implemented.
  contact:
    name: TribeHub Team

servers:
  - url: http://localhost:3000/api/v1
    description: Local development

security:
  - BearerAuth: []

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      tags:
        - Health
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '500':
          description: Service is unhealthy or an internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/register:
    post:
      operationId: registerUser
      summary: Register a new user account
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully. Email confirmation may be required before login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      operationId: loginUser
      summary: Authenticate with email and password
      tags:
        - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials or email not confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /profile/me:
    get:
      operationId: getMyProfile
      summary: Get current user profile
      tags:
        - Profile
      responses:
        '200':
          description: Profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateMyProfile
      summary: Update current user profile
      tags:
        - Profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate the refresh token and obtain a new access token
      tags:
        - Auth
      security:
        - RefreshTokenCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
          description: CSRF token value from XSRF-TOKEN cookie
      responses:
        '200':
          description: Tokens rotated successfully. Sets new refresh cookie and XSRF cookie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                  message:
                    type: string
        '401':
          description: Refresh token invalid, expired, or blacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      operationId: logoutUser
      summary: Revoke the current session and blacklist the refresh token
      tags:
        - Auth
      security:
        - RefreshTokenCookie: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Session revoked successfully
        '401':
          description: Missing or invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: supersecret123

    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: Registration successful. Please check your email to confirm your account.

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: supersecret123

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: 'Short-lived JWT access token (Authorization: Bearer)'
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. INVALID_CREDENTIALS, USER_NOT_FOUND)
          example: INVALID_CREDENTIALS
        message:
          type: string
          description: Human-readable error description
          example: The provided credentials are incorrect
        details:
          type: array
          description: Optional list of field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Profile ID
          example: 550e8400-e29b-41d4-a716-446655440000
        userId:
          type: string
          format: uuid
          description: User ID
          example: 550e8400-e29b-41d4-a716-446655440000
        displayName:
          type: string
          description: Display name (nombre visible)
          example: Alex García
        bio:
          type: string
          description: Short bio
          nullable: true
          example: Hola, soy Alex!
        avatarUrl:
          type: string
          description: Avatar URL
          nullable: true
          example: https://cdn.tribehub.app/avatars/user_123/avatar.png
        isPublic:
          type: boolean
          description: Whether profile is public
          example: true
        createdAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-01T00:00:00.000Z

    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 50
          description: Display name (nombre visible)
          example: Alex García
        bio:
          type: string
          maxLength: 160
          description: Short bio (optional)
          nullable: true
          example: Hola, soy Alex!
        avatarUrl:
          type: string
          format: uri
          description: Avatar URL (optional)
          nullable: true
          example: https://cdn.tribehub.app/avatars/user_123/avatar.png